+++ src/main/java/blackjack/domain/Deck.java
@@ -1,5 +1,9 @@
 package blackjack.domain;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardNumber;
+import blackjack.domain.card.Shape;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
