+++ src/main/java/blackjack/domain/GameResult.java
@@ -1,5 +1,10 @@
 package blackjack.domain;
 
+import blackjack.domain.card.GamePoint;
+import blackjack.domain.user.Dealer;
+import blackjack.domain.user.Name;
+import blackjack.domain.user.User;
+
 import java.util.*;
 
 public class GameResult {
