+++ src/main/java/blackjack/domain/Users.java
@@ -1,5 +1,10 @@
 package blackjack.domain;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.GamePoint;
+import blackjack.domain.user.Name;
+import blackjack.domain.user.User;
+
 import java.util.List;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
