+++ src/main/java/blackjack/domain/card/Card.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import java.util.Objects;
