+++ src/main/java/blackjack/domain/card/GamePoint.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import java.util.List;
