+++ src/main/java/blackjack/domain/user/Dealer.java
@@ -1,4 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.user;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Cards;
+import blackjack.domain.card.GamePoint;
 
 import java.util.ArrayList;
 import java.util.List;
