+++ src/test/java/blackjack/domain/CardNumberTest.java
@@ -1,5 +1,6 @@
 package blackjack.domain;
 
+import blackjack.domain.card.CardNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
