+++ src/test/java/blackjack/domain/CardsTest.java
@@ -1,6 +1,9 @@
 package blackjack.domain;
 
-import org.assertj.core.api.Assertions;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardNumber;
+import blackjack.domain.card.Cards;
+import blackjack.domain.card.Shape;
 import org.assertj.core.api.InstanceOfAssertFactories;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
