+++ src/test/java/blackjack/domain/UserTest.java
@@ -1,5 +1,11 @@
 package blackjack.domain;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardNumber;
+import blackjack.domain.card.Cards;
+import blackjack.domain.card.Shape;
+import blackjack.domain.user.Name;
+import blackjack.domain.user.User;
 import org.assertj.core.api.InstanceOfAssertFactories;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
